Update dispatcher.lua
added 3 new functions for new multi-user functionality. Rewrote the entry function to check permissions instead of performing the check in the controllers. This should provide all available menus to be enabled disabled ... note we still can't handle page/nodes so the must still be evaluated inside the controller .. hopefully i can tap into to the node plumbing in the next release.